💄 Design: 게스트 헤더 공통 스타일로 적용하여 수정
